requirement.txt
astroid==2.0.3
beautifulsoup4==4.6.0
certifi==2018.4.16
chardet==3.0.4
Django==2.0.8
django-ckeditor==5.6.1
django-crispy-forms==1.7.2
django-filter==2.0.0
django-js-asset==1.1.0
django-modelcluster==4.2
django-taggit==0.23.0
django-treebeard==4.3
djangorestframework==3.8.2
draftjs-exporter==2.1.2
html5lib==1.0.1
idna==2.7
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
Pillow==5.2.0
pkg-resources==0.0.0
pylint==2.1.1
pytz==2018.5
requests==2.19.1
six==1.11.0
taggit-selectize==2.5.0
typed-ast==1.1.0
Unidecode==0.4.21
urllib3==1.23
webencodings==0.5.1
Willow==1.1
wrapt==1.10.11
